+++ src/org/trifort/rootbeer/testcases/rootbeertest/kerneltemplate/DoubleToStringKernelTemplateBuilderTest.java
@@ -21,7 +21,7 @@ public class DoubleToStringKernelTemplateBuilderTest implements TestKernelTempla
   
   @Override
   public ThreadConfig getThreadConfig() {
-    return new ThreadConfig(m_blockSize, m_gridSize, m_blockSize * m_gridSize);
+    return new ThreadConfig(m_blockSize, m_gridSize, (long) m_blockSize * (long) m_gridSize);
   }
   
   @Override
