+++ src/org/trifort/rootbeer/testcases/rootbeertest/kerneltemplate/FastMatrixTest.java
@@ -38,7 +38,7 @@ public class FastMatrixTest implements TestKernelTemplate {
   }
 
   public ThreadConfig getThreadConfig() {
-    ThreadConfig ret = new ThreadConfig(m_blockSize, m_gridSize,m_blockSize * m_gridSize);
+    ThreadConfig ret = new ThreadConfig(m_blockSize, m_gridSize, (long) m_blockSize * (long) m_gridSize);
     return ret;
   }
